+++ src/algo/ImportBlastHits.cpp
@@ -65,7 +65,7 @@ struct BlastHit {
         }
         items[0].id = parts[0];
         items[1].id = parts[1];
-        ident = boost::lexical_cast<float>(parts[2]) / 100;
+        ident = boost::lexical_cast<double>(parts[2]) / 100;
         length = boost::lexical_cast<int>(parts[3]);
         mismatches = boost::lexical_cast<int>(parts[4]);
         gap_openings = boost::lexical_cast<int>(parts[5]);
@@ -78,7 +78,7 @@ struct BlastHit {
     }
 
     BlastItem items[2];
-    float ident;
+    double ident;
     int length;
     int mismatches;
     int gap_openings;
