+++ src/algo/Stats.cpp
@@ -40,8 +40,8 @@ static void report_part(std::ostream& o, const std::string& name,
                         T1 part, T2 total) {
     o << name << ": " << part;
     if (total) {
-        float portion = float(part) / float(total);
-        float percentage = portion * 100;
+        double portion = float(part) / float(total);
+        double percentage = portion * 100;
         o << " (" << percentage << "%)";
     }
     o << "\n";
@@ -103,7 +103,7 @@ void Stats::run_impl() const {
     ASSERT_GTE(total_seq_length, unique_nucl);
     size_t seq_nucl_in_blocks = total_seq_length - unique_nucl;
     size_t bss = block_set()->size();
-    float fpb = bss ? float(total_fragments) / bss : 0;
+    double fpb = bss ? float(total_fragments) / bss : 0;
     std::ostream& out = file_writer_.output();
     if (total_fragments != bss) {
         out << "fragments / blocks = ";
