+++ src/model/Block.hpp
@@ -115,7 +115,7 @@ public:
     If a fragment doesn't have alignment row attached,
     then it is taken as is.
     */
-    float identity() const;
+    double identity() const;
 
     /** Return consensus letter for given position.
     For each column, the most frequent letter is written to consensus.
